improve file organisation, improve error logging
