+++ orders/chalicelib/src/constants.py
@@ -3,10 +3,10 @@ from decimal import Decimal
 import pytz
 
 # Developer Constants
-development_mode = False
+development_mode: bool = False
 
 # General finance constants
-country_personal_tax_rate: Decimal = 0.42  # "<insert> tax rate"
+country_personal_tax_rate: Decimal = 0.42
 
 tax_rate: Decimal = country_personal_tax_rate
 
@@ -16,4 +16,4 @@ capital_to_deploy_percentage: Decimal = 0.33
 tradingview_ticker: str = "ticker"
 
 # Timezone constants
-berlin_tz = pytz.timezone("Europe/Berlin")
+local_tz = pytz.timezone("Europe/Berlin")
